Выполнил вторую часть ДЗ к лекции №5: "Структурные шаблоны". Реализован паттерн Adapter для связи внешней библиотеки с имеющимся кодом. Для устранения огромной погрешности пришлось изменить типы данных в интерфейсах с int на float.

+++ Homework/Lesson5/task2/CircleAreaLib.php
@@ -2,7 +2,7 @@
 
 class CircleAreaLib
 {
-    public function getCircleArea(int $diagonal)
+    public function getCircleArea(float $diagonal)
     {
         $area = (M_PI * $diagonal ** 2) / 4;
         return $area;
